Writing side is now according to spec
