+++ src/create_tx.rs
@@ -58,6 +58,60 @@ impl CreateTx {
 		self.writer.add_record(key, format, data)
 	}
 
+    pub fn delete(
+        &mut self,
+        first_key: &str,
+        last_key: &str,
+        before_time: u64,
+        after_time: u64,
+        filter: &str,
+    ) -> std::result::Result<(), crate::write::WriteFailure> {
+        use crate::row_format::{
+            Element as _,
+            ElementString,
+        };
+        use byteorder::{BigEndian, ByteOrder as _};
+
+        // write row format
+        let key = first_key;
+        let format = "\u{00f7}";
+
+        let mut row_data = Vec::with_capacity(4);
+        /*
+            unimplemented!() // last timestamp length
+            + unimplemented!() // key wildcard length
+            + unimplemented!() // last key length
+        );
+        */
+        let cap = row_data.capacity();
+
+        // bypass RowFormat entirely, we're going to be building row_data here so
+        // we don't get to pass str values
+
+        // write first key
+        ElementString.to_stored_format(&first_key, &mut row_data);
+        dbg!(row_data.len());
+        
+        // write first timestamp
+        BigEndian::write_u64(&mut row_data, before_time);
+        dbg!(row_data.len());
+        
+        // write last timestamp
+        BigEndian::write_u64(&mut row_data, after_time);
+        dbg!(row_data.len());
+        
+        // write key wildcard
+        ElementString.to_stored_format(&filter, &mut row_data);
+        dbg!(row_data.len());
+        
+        // write last key
+        ElementString.to_stored_format(&last_key, &mut row_data);
+
+        assert!(cap == row_data.capacity());
+
+        self.add_record(&key, &format, &row_data)
+    }
+
 	/// Commit the transaction, but give it a specific name.
 	///
 	/// This function is necessary for compacting, normally
